requirements.txt
apache-airflow==2.7.2
pandas==2.0.3
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9